\name{hanna-package} 
\alias{hanna} 
\alias{hanna-package} 
\title{Package hanna - simulations for winner-looser effects} 
\description{This package contains the code for the paper  
  "Hermanussen et. al. (2023): Winner-loser effects improve social network efficiency among  
   competitors with equal resource holding power." 
 
 The following objects / modules are available: 
\describe{ 
  \item{\link[hanna:hgraph]{hgraph}}{ R utility code functions for working with graphs} 
  \item{\link[hanna:simul]{simul}}{ R code for the simulation of the effect} 
} 
} 
 
\author{Detlef Groth and Sarah Leutner, University of Potsdam, Germany} 
 
