\name{simul}  
\alias{simul}  
\title{ Collection of functions which do winner-looser and network structure simulations. }  
\description{  
  The function with the simul prefix perform simulations of winner-looser effects for the paper ... 
  The following functions are implemented: 
  \itemize{ 
     \item \code{\link{simul_pairings}} - create round pairings for a season 
   } 
} 
 
