\name{simul}  
\alias{simul}  
\title{ Collection of functions which do winner-looser and network structure simulations. }  
\description{  
  The function with the simul prefix perform simulations of winner-looser effects for the paper ... 
  The following functions are implemented: 
  \itemize{ 
     \item \code{\link{simul_pairings}} - create round pairings for a season 
     \item \code{\link{simul_season}} - reate matches for everyone against everyone using the given model 
     \item \code{\link{simul_graph}} - create a adjacency matrix out of the results for a match season 
     \item \code{\link{simul_plot}} - plot an adjacency matrix  
     \item \code{\link{simul_layout}} - calculate a plot layout for an adjacency matrix 
     \item \code{\link{simul_colors}} - create a color vector based on node degrees  
     \item \code{\link{simul_compare}} - compare the different models for a certain number of seasons 
     \item \code{\link{simul_triads}} - alculate the number of two and tri edge triads 
   } 
} 
 
